Basic Table done, starting customer cell
